Have no_log be smarter
The NoLog struct field maps _almost_ to the Sensitive field from the
upstream mmv1 repo. The difference is that ansible's sanity tests
perform some extra heuristics to check if a field *looks like* it should
be a no_log field or not.
In order to quiet the sanity check failure, you have to explicitly set
no_log to false, so in the case of a field that looks like a secret but
isn't a secret, we set that thanks to NoLog being a pointer field now
which means it accepts a nil value on top of yes/no bit